+++ GameParticles/ParticleEmitter.h
@@ -24,22 +24,22 @@ public:
 
 private:
 
-  const Vect4D	start_position = Vect4D(0.0, 0.0, 0.0);
-  const Vect4D	start_velocity = Vect4D(0.0, 1.0, 0.0);
+  const Vect4D	start_position = Vect4D(0.0f, 0.0f, 0.0f);
+  const Vect4D	start_velocity = Vect4D(0.0f, 1.0f, 0.0f);
 
-  const double	max_life = MAX_LIFE;
+  const float	max_life = MAX_LIFE;
   const int		max_particles = NUM_PARTICLES;
-  const double	spawn_frequency = 0.0000001;
+  const float	spawn_frequency = 0.0000001f;
 
-  double	last_spawn;
-  double	last_loop;
+  float	last_spawn;
+  float	last_loop;
   int		last_active_particle;
 
   // added for speed efficiency
   int bufferCount;
   Vect4D	vel_variance;
   Vect4D	pos_variance;
-  const double	scale_variance = 2.5;
+  const float	scale_variance = 2.5f;
   std::list<Particle> drawBuffer;
   Particle* headParticle;
